+++ packages/admin/admin-date-time/CHANGELOG.md
@@ -1,6 +1,5 @@
 # @comet/admin-date-time
 
-<<<<<<< HEAD
 ## 9.0.0-beta.0
 
 ### Major Changes
@@ -57,7 +56,7 @@
 - Updated dependencies [631540c]
     - @comet/admin-icons@9.0.0-beta.0
     - @comet/admin@9.0.0-beta.0
-=======
+
 ## 8.20.0
 
 ### Patch Changes
@@ -84,7 +83,6 @@
 - Updated dependencies [d344f53]
     - @comet/admin@8.18.0
     - @comet/admin-icons@8.18.0
->>>>>>> main
 
 ## 8.17.1